feat: Implement base entity and context for Email API with PostgreSQL and SQL Server support
