+++ Mango.Common.Extensions/Data/BaseAppDbContext.cs
@@ -1,0 +1,42 @@
+using Mango.Common.Extensions.Models;
+using Microsoft.EntityFrameworkCore;
+
+namespace Mango.Common.Extensions.Data
+{
+    public abstract class BaseAppDbContext : DbContext
+    {
+        protected BaseAppDbContext(DbContextOptions options) : base(options)
+        {
+        }
+
+        public override int SaveChanges()
+        {
+            UpdateTimestamps();
+            return base.SaveChanges();
+        }
+
+        public override async Task<int> SaveChangesAsync(CancellationToken cancellationToken = default)
+        {
+            UpdateTimestamps();
+            return await base.SaveChangesAsync(cancellationToken);
+        }
+
+        private void UpdateTimestamps()
+        {
+            var entries = ChangeTracker.Entries<BaseEntity>();
+
+            foreach (var entry in entries)
+            {
+                if (entry.State == EntityState.Added)
+                {
+                    entry.Entity.CreatedAt = DateTime.Now;
+                    entry.Entity.UpdatedAt = DateTime.Now;
+                }
+                if (entry.State == EntityState.Modified)
+                {
+                    entry.Entity.UpdatedAt = DateTime.Now;
+                }
+            }
+        }
+    }
+}
