+++ Mango.Common.Extensions/Models/BaseEntity.cs
@@ -1,6 +1,6 @@
-﻿namespace Mango.Services.EmailAPI.Models
+namespace Mango.Common.Extensions.Models
 {
-    public class BaseEntity
+    public abstract class BaseEntity
     {
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
