+++ Mango.Services.EmailAPI/Data/AppDbContext.cs
@@ -1,41 +1,15 @@
-﻿using Mango.Services.EmailAPI.Models;
+﻿using Mango.Common.Extensions.Data;
+using Mango.Services.EmailAPI.Models;
 using Microsoft.EntityFrameworkCore;
 
 namespace Mango.Services.EmailAPI.Data
 {
-    public class AppDbContext(DbContextOptions<AppDbContext> options) : DbContext(options)
+    public class AppDbContext : BaseAppDbContext
     {
+        public AppDbContext(DbContextOptions options) : base(options)
+        {
+        }
+
         public DbSet<EmailLogger> EmailLoggers { get; set; }
-
-        public override int SaveChanges()
-        {
-            UpdateTimestamps();
-            return base.SaveChanges();
-        }
-
-        public override async Task<int> SaveChangesAsync(CancellationToken cancellationToken = default)
-        {
-            UpdateTimestamps();
-            return await base.SaveChangesAsync(cancellationToken);
-
-        }
-
-        private void UpdateTimestamps()
-        {
-            var entries = ChangeTracker.Entries<BaseEntity>();
-
-            foreach (var entry in entries)
-            {
-                if (entry.State == EntityState.Added)
-                {
-                    entry.Entity.CreatedAt = DateTime.Now;
-                    entry.Entity.UpdatedAt = DateTime.Now;
-                }
-                if (entry.State == EntityState.Modified)
-                {
-                    entry.Entity.UpdatedAt = DateTime.Now;
-                }
-            }
-        }
     }
 }
