+++ Mango.Services.EmailAPI/Data/PostgreSqlAppDbContext.cs
@@ -1,0 +1,11 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace Mango.Services.EmailAPI.Data
+{
+    public class PostgreSqlAppDbContext : AppDbContext
+    {
+        public PostgreSqlAppDbContext(DbContextOptions<PostgreSqlAppDbContext> options) : base(options)
+        {
+        }
+    }
+}
