+++ Mango.Services.EmailAPI/Data/SqlServerAppDbContext.cs
@@ -1,0 +1,11 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace Mango.Services.EmailAPI.Data
+{
+    public class SqlServerAppDbContext : AppDbContext
+    {
+        public SqlServerAppDbContext(DbContextOptions<SqlServerAppDbContext> options) : base(options)
+        {
+        }
+    }
+}
